Move most types from shared crate
It only needs to contain interface definitions
since they are used by the derive crate.
