+++ crates/rt/src/data/persistence.rs
@@ -14,8 +14,11 @@ use super::{
     time::Timers,
     PersistedWorkflowData, WorkflowCounters, WorkflowData,
 };
-use crate::{module::Services, workflow::ChannelIds, WorkflowId};
-use tardigrade::interface::{ChannelKind, Interface};
+use crate::{module::Services, workflow::ChannelIds};
+use tardigrade::{
+    interface::{ChannelKind, Interface},
+    WorkflowId,
+};
 
 /// Error persisting a workflow.
 #[derive(Debug)]
