+++ src/test.rs
@@ -106,9 +106,9 @@ use crate::{
     interface::Interface,
     spawn::{ManageWorkflowsExt, RemoteWorkflow, Spawner, Workflows},
     task::{self, TaskResult},
+    types::WorkflowId,
     workflow::{Handle, SpawnWorkflow, TakeHandle, TaskHandle, UntypedHandle, Wasm},
 };
-use tardigrade_shared::WorkflowId;
 
 #[derive(Debug)]
 struct TimerEntry {
